pyproject.toml
[tool.poetry]
name = "moshimoshi"
version = "0.1.5"
description = "call a function in any python scripts using a formated string or json"
authors = ["Nutchanon Ninyawee <me@nutchanon.org>"]
license = "ISC"
readme = "README.md"
repository = "https://github.com/CircleOnCircles/moshimoshi"

[tool.poetry.dependencies]
python = "^3.7"
parse_with_dot_access = "^1.17.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"
pytest = "^6.0.2"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
